move highlight new file
